+++ packages/shopify-app-remix/src/server/authenticate/admin/__tests__/session-token-header-path.test.ts
@@ -30,59 +30,6 @@ describe('authorize.session token header path', () => {
       // THEN
       expect(response.status).toBe(401);
     });
-
-    describe.each([true, false])('when isOnline: %s', (isOnline) => {
-      it(`returns app bridge redirection headers if there is no session`, async () => {
-        // GIVEN
-        const shopify = shopifyApp(testConfig({useOnlineTokens: isOnline}));
-
-        // WHEN
-        const {token} = getJwt();
-        const response = await getThrownResponse(
-          shopify.authenticate.admin,
-          new Request(`${APP_URL}?shop=${TEST_SHOP}&host=${BASE64_HOST}`, {
-            headers: {Authorization: `Bearer ${token}`},
-          }),
-        );
-
-        // THEN
-        const {origin, pathname, searchParams} = new URL(
-          response.headers.get(REAUTH_URL_HEADER)!,
-        );
-
-        expect(response.status).toBe(401);
-        expect(origin).toBe(APP_URL);
-        expect(pathname).toBe('/auth');
-        expect(searchParams.get('shop')).toBe(TEST_SHOP);
-      });
-
-      it(`returns app bridge redirection headers if the session is no longer valid`, async () => {
-        // GIVEN
-        const shopify = shopifyApp(
-          testConfig({useOnlineTokens: isOnline, scopes: ['otherTestScope']}),
-        );
-        // The session scopes don't match the configured scopes, so it needs to be reset
-        await setUpValidSession(shopify.sessionStorage, isOnline);
-
-        // WHEN
-        const {token} = getJwt();
-        const response = await getThrownResponse(
-          shopify.authenticate.admin,
-          new Request(`${APP_URL}?shop=${TEST_SHOP}&host=${BASE64_HOST}`, {
-            headers: {Authorization: `Bearer ${token}`},
-          }),
-        );
-
-        // THEN
-        const {origin, pathname, searchParams} = new URL(
-          response.headers.get(REAUTH_URL_HEADER)!,
-        );
-        expect(response.status).toBe(401);
-        expect(origin).toBe(APP_URL);
-        expect(pathname).toBe('/auth');
-        expect(searchParams.get('shop')).toBe(TEST_SHOP);
-      });
-    });
   });
 
   describe.each([true, false])(
@@ -92,10 +39,9 @@ describe('authorize.session token header path', () => {
         // GIVEN
         const shopify = shopifyApp(testConfig({useOnlineTokens: isOnline}));
 
-        const testSession = await setUpValidSession(
-          shopify.sessionStorage,
+        const testSession = await setUpValidSession(shopify.sessionStorage, {
           isOnline,
-        );
+        });
 
         // WHEN
         const {token, payload} = getJwt();
@@ -120,7 +66,9 @@ describe('authorize.session token header path', () => {
         let testSession: Session;
         testSession = await setUpValidSession(shopify.sessionStorage);
         if (isOnline) {
-          testSession = await setUpValidSession(shopify.sessionStorage, true);
+          testSession = await setUpValidSession(shopify.sessionStorage, {
+            isOnline: true,
+          });
         }
 
         // WHEN