+++ packages/shopify-app-remix/src/server/unauthenticated/admin/__tests__/factory.test.ts
@@ -17,13 +17,11 @@ describe('unauthenticated admin context', () => {
 
   expectAdminApiClient(async () => {
     const shopify = shopifyApp(testConfig());
-    const expectedSession = await setUpValidSession(
-      shopify.sessionStorage,
-      false,
-    );
-    const {admin, session: actualSession} = await shopify.unauthenticated.admin(
-      TEST_SHOP,
-    );
+    const expectedSession = await setUpValidSession(shopify.sessionStorage, {
+      isOnline: false,
+    });
+    const {admin, session: actualSession} =
+      await shopify.unauthenticated.admin(TEST_SHOP);
 
     return {admin, expectedSession, actualSession};
   });