+++ packages/shopify-app-session-storage-sqlite/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @shopify/shopify-app-session-storage-sqlite
 
+## 2.0.4
+
+### Patch Changes
+
+- b4eeb24: Improved and simplified package.json dependencies
+- b998c30: Bump shopify-api version from 9.0.1 to 9.0.2
+- Updated dependencies [b4eeb24]
+- Updated dependencies [b998c30]
+  - @shopify/shopify-app-session-storage@2.0.4
+
 ## 2.0.3
 
 ### Patch Changes